SolitaireSolver: Add model Card
This adds an abstract class Card, which defines some constants for
describing a card as an integer, as well a function for generating a
full deck of cards.

A card is represented as an integer, where the lowest four bits describe
the rank of the card, and the following two bits describe the suit. A
rank of 13 means that the card is unknown/facedown.