SolitaireSolver: Add record Klondike
This adds the record Klondike and the classes Foundation, Column, and
Stock, representing respectively a game of Klondike and the components
hereof. These are far from finished, but it is a start at least.

Right now, Foundation, Column, and Stock each extend a collection. I
do not think that this is likely to make sense long-term, but it will
have to do for now.