SolitaireSolver: Add Klondike::Column::reveal(card, index)
This adds the method Column::reveal(card, index) which sets a
Card::Unknown at the specified index to the specified card, or, if there
is no such card, throws an IllegalArgumentException.

This is not an ideal solution; right now there is nothing to ensure that
a game couldn't end up with duplicate cards. It is possible that it thus
should be the responsibility of Klondike::Klondike to reveal cards in
its columns, or that a Column somehow should know what game it is part
of, but both of these options, especially the latter, seems rather icky.

Perhaps a Column could keep track of available cards, and Klondike could
just subscribe to its columns and update the others when a card is used,
but this will result in a bunch of duplicated information and
unnecessary overhead. Maybe some sort of Deck class could be introduced,
a single instance of which could shared by each Column?