SolitaireSolver: Add method Solitaire::possibleMoves
This adds an interface Solitaire, which currently only specifies one
method, Solitaire::possibleMoves, that returns a List<Move>. The newly
added dataclass Move consists of an Integer representing a Card to be
moved and an Optional<Integer> representing the destination Card. In the
case of an Ace being moved to an empty Foundation, or a King being moved
to an empty Column, the destination card will be Optional::empty.

Klondike, now implementing Solitaire, of course
implements this method, though the implementation is as of yet
incomplete. Currently, it returns only moves from the columns to the
foundations, and is in dire need of refactoring, but it is at least a
start :^)