SolitaireSolver: Add method klondike::Stock::take(card)
This adds the method klondike::Stock::take, which, if possible, takes
the specified card from the klondike::Stock. This method will fail if
the specified card either is not contained in the klondike::Stock or if
it is not reachable.

This is not currently reversible, which likely will be necessary for
generating different game states when solving the game, but it will do
for now.

It also is pretty icky, being quite far from optimal, but one must be
wary of premature optimisations :^)