SolitaireSolver: Change Klondike::findMoveSource to Klondike::findSource
Klondike::findMoveSource(Move) did not actually make use of the move,
and only looked for the moving card. This simplifies the method, makes
it more general, useful for shortly when moves need to be undone, and as
a bonus, also means its argument is passed by value rather than
reference which hopefully is ever so slightly faster.