SolitaireSolver: Add method CardContainer::undo
This adds, to the interface CardContainer, the method
CardContainer::undo(card, cardDestination, moveMetaInformation), which
will undo a move.

It would likely be a good idea to include cardDestination and card in
moveMetaInformation, and pass it all as a single parameter.

Implementations have been provided for klondike::Column and
klondike::Foundation, but not yet for klondike::Stock. :^)