SolitaireSolver: Add method Solitaire::undoMove()
This adds the method Solitaire::undoMove() and implements it in the
class klondike::Klondike. This is done by using a stack to keep track
of the move history.

klondike::Klondike now also has a klondike::Klondike::equals,
klondike::Klondike::deepCopy and klondike::Klondike::hashCode. The first
two are mostly useful in testing, the latter however may be used in the
future to hash the game state, and using such hashes to keep track of
whether a state has been seen before or not.