SolitaireSolver: Refactor klondike::Column
This refactors klondike::Column such that it no longer extends vector.
This means that it is now protected from unwanted meddling in its
internal state, and much more easily can ensure that it is always valid.

It of course still is not optimised, but at least a lot cleaner and
easier to work with. :^)