SolitaireSolver: Add function Card::validateCard
This adds the static method Card::validateCard which simply calls the
function Card::isValidCard and throws an InvalidCardException if this
returns false.

This also adds the class InvalidCardException, which, as the name
implies should be used when an integer is expected to be a Card, but is
not. InvalidCardException currently extends IllegalArgumentException,
however this might be slightly changed in the future.