SolitaireSolver: Add function Card::isUnknown(int card)
This adds the static method Card::isUnknown(int card), which returns a
boolean indicating whether the specified card is of unknown rank. It has
turned out to be rather easy to forget that an unknown card technically
can have a suit. This function ensures that this is accounted for, when
checking if a card is unknown.